:bug: Fix #15, bad anonymous function assignment in API stores
The API store JS file assigned the WP function `registerStore` to
an anonymous function `registerStores` assigned to an exported
constant. The Firefox dev tools threw an exception when the main
script executed that constant as a function; arguing that it was
not a function. The page loaded and functioned correctly, but to
avoid the error we recreated the anonymous function as a standard
function containing `registerStore`.